Lots of changes to support NETCore, added NETCore test project
Stopped using reflection to call the ConvertToElement() methods since
NETCore complained about it. Changed some tests so that they're in a
shared project (the Portable library can't reference the shared project
though, so had to use linking still for that). Changed tile tests so
that they test the public API, not the internal classes.